feat: Add animated underline for links
Introduces a new CSS class `animated-underline` to provide a subtle, animated underline effect on hover for links. This enhances the visual feedback and user experience. The "About" link in the header has been replaced with a "合租" link that uses this new animation. The "About" link has been moved to the footer.